+++ CHANGELOG.md
@@ -4,6 +4,13 @@
 格式基於 [Keep a Changelog](http://keepachangelog.com/zh-TW/1.0.0/)
 這個項目堅持 [Semantic Versioning](http://semver.org/spec/v2.0.0.html)
 
+## [0.0.3] - 2018-08-16
+### Added
+- .gitignore 增加路徑 envs/, credentials/
+- 新增 tools 資料夾
+- 在 tools 資料夾下加入 analyzer, bilibili 的 package
+- 新增 requirement.txt
+
 ## [0.0.2] - 2018-08-10
 ### Added
 - 新增 helper 資料夾